+++ packages/jars/lib/bloc_utilis.dart
@@ -1,5 +1,9 @@
 library bloc_utils;
 
-export 'src/bloc_utils/bloc_event_transformers.dart';
-export 'src/bloc_utils/bloc_observer.dart';
+export 'src/bloc_utils/event_transformer/bloc_event_transformers.dart';
+export 'src/bloc_utils/observer/bloc_observer.dart';
+export 'src/bloc_utils/replay/replay_cubit.dart';
 export 'src/bloc_utils/bloc_state.dart';
+
+export 'src/bloc_utils/bloc_selective_consumer/bloc_selective_consumer.dart';
+export 'src/bloc_utils/bloc_selective_consumer/droppable_mixin.dart';